refactor(lib): move ResetPassword to Login and remove transition
