+++ packages/lib/src/components/common/Login/components/LoginForm/index.ts
@@ -1,1 +1,2 @@
+export * from './components'
 export * from './LoginForm'
