--- packages/lib/src/components/common/ResetPassword/index.ts
@@ -1,1 +1,0 @@
-export * from './components'
