+++ packages/lib/src/components/common/index.ts
@@ -3,5 +3,4 @@ export * from './Footer'
 export * from './Header'
 export * from './Login'
 export * from './NavBar'
-export * from './ResetPassword'
 export * from './SetPassword'
